+++ app/views/schools/school_targets/progress/_charts.html.erb
