+++ app/views/schools/school_targets/progress/_row.html.erb
