+++ app/views/schools/school_targets/progress/index.html.erb
