+++ config/routes.rb
@@ -146,7 +146,19 @@ Rails.application.routes.draw do
           get :storage_heater
         end
       end
-      resources :school_targets
+
+      resources :school_targets do
+        scope module: :school_targets do
+          resources :progress do
+            collection do
+              get :electricity
+              get :gas
+              get :storage_heater
+            end
+          end
+        end
+      end
+
       resources :estimated_annual_consumptions, except: [:show]
 
       resources :programmes, only: [:create]
